+++ p2panda-stream/src/groups/args.rs
@@ -4,10 +4,12 @@
 use p2panda_core::{Hash, VerifyingKey};
 use serde::{Deserialize, Serialize};
 
-use crate::group::GroupAction;
+use p2panda_auth::group::GroupAction;
 
 /// Additional arguments which can be attached to a p2panda operation in their extensions.
 #[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
+// TODO(glyph): Consider renaming to `GroupsExtensionArgs` to further differentiate from
+// `GroupsProcessorArgs`.
 pub struct GroupsArgs<C = ()> {
     pub group_id: VerifyingKey,
     pub action: GroupAction<VerifyingKey, C>,